pyproject.toml
[project]
name = "flash-post"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.2",
    "fastapi[standard]>=0.116.0",
    "psycopg>=3.2.9",
    "pydantic-settings>=2.10.1",
    "sqlalchemy>=2.0.41",
    "asyncpg>=0.30.0",
    "passlib>=1.7.4",
]
